devlog is now shortlog; alias placed in legacy; bower moved to legacy

+++ Task/Log/Short.php
@@ -7,15 +7,22 @@
  * @copyright  (c) 2013, Ibidem Team
  * @license    https://github.com/ibidem/ibidem/blob/master/LICENSE.md
  */
-class Task_Devlog extends \app\Task_Base
+class Task_Log_Short extends \app\Task_Base
 {
 	/**
 	 * Execute task.
 	 */
 	function run()
 	{
+		$erase = $this->get('erase', false);
+
 		$logpath = \app\Env::key('etc.path').'logs/short.log';
-		\file_put_contents($logpath, '');
+
+		if ($erase || !\file_exists($logpath))
+		{
+			\file_put_contents($logpath, '');
+		}
+
 		\passthru("tail -f $logpath");
 	}
